[fix] int -> Long 타입 변환

+++ TradingMatchingService/src/main/java/com/rktpdyfk/TradingMatchingService/dto/OptionPostDto.java
@@ -4,17 +4,17 @@ import jakarta.persistence.Column;
 
 public class OptionPostDto {
     public static class OptionPostRequestDto {
-        private int weaponAttack;
-        private int magicAttack;
-        private int upgradeSlots;
-        private int str;
-        private int dex;
-        private int intell;
-        private int luk;
-        private int weaponDef;
-        private int magicDef;
-        private int speed;
-        private int jump;
-        private int dodge;
+        private Long weaponAttack;
+        private Long magicAttack;
+        private Long upgradeSlots;
+        private Long str;
+        private Long dex;
+        private Long intell;
+        private Long luk;
+        private Long weaponDef;
+        private Long magicDef;
+        private Long speed;
+        private Long jump;
+        private Long dodge;
     }
 }
